### 2023-03-20 #5271 @editorialbot
[PRE REVIEW]: gesel: a JavaScript package for client-side gene set enrichment
**Submitting author:** @LTLA (<a href="http://orcid.org/0000-0002-3564-4813">Aaron Lun</a>)
**Repository:** https://github.com/LTLA/gesel.js
**Branch with paper.md** (empty if default branch): submission
**Version:** 0.2.3
