### 2023-03-20 4ff3c81
Added reindexing utilities for easier mapping to user universes.
This is intended for applications that already know their full gene
universe and want to transform the existing set<->gene mappings in the
gesel database to work with gene indices into their own universe.
