### 2023-03-20 f5a5545
Factored out the set overlap counter for easier re-use.
