### 2023-03-20 996af00
Use a dual-purpose getter/setter for the download functions.
This makes life easier for applications that want to use the existing
downloaders but also want to apply some customizations on top, e.g., for
caching or editing of the headers.
