### 2023-03-21 b8aa1e1
Re-added code to compute BH-adjusted p-values.
This allows users to control the FDR across gene sets, as unwise as that idea
may be. In practice, the p-values from findOverlappingSets will need to be
adjusted with an appropriate setting of totalTests to account for the true
amount of multiple testing that was implicitly done during the search.
