### 2023-03-22 10494e9
Use Ramanujan's approximation for more accurate log-factorials.
