### 2023-03-22 2b0d368
Added utility to compute enrichment curves from gene rankings.
This uses a proportion-based metric to quantify enrichment, which is a more
intuitive measure of enrichment compared to the subtraction method used by the
original GSEA software. We apply a pseudocount to stabilize the curve's start.
