+++ src/index.js
@@ -19,6 +19,7 @@ export { findOverlappingSets, countSetOverlaps } from "./findOverlappingSets.js"
 
 export { testEnrichment } from "./testEnrichment.js";
 export { adjustFdr } from "./adjustFdr.js";
+export { computeEnrichmentCurve } from "./computeEnrichmentCurve.js";
 
 export { reindexGenesForAllSets } from "./reindexGenesForAllSets.js";
 export { reindexSetsForAllGenes } from "./reindexSetsForAllGenes.js";
