### 2023-03-24 7a122f0
Fixed documentation for the reindexing functions.
