### 2023-04-29 #5271 @Kevin-Mattheus-Moerman
@LTLA for moment would you be able to work on those potentially missing DOIs ☝️ You can fix them in your .bib file and then call `@editorialbot check references` here to check them again. You can also call `@editorialbot generate pdf` to update the paper draft. Thanks.
