+++ ref.bib
@@ -2,6 +2,7 @@
    Author="Chen, E. Y.  and Tan, C. M.  and Kou, Y.  and Duan, Q.  and Wang, Z.  and Meirelles, G. V.  and Clark, N. R.  and Ma'ayan, A. ",
    Title="{{E}nrichr: interactive and collaborative {H}{T}{M}{L}5 gene list enrichment analysis tool}",
    Journal="BMC Bioinformatics",
+   doi="10.1186/1471-2105-14-128",
    Year="2013",
    Volume="14",
    Pages="128",
@@ -36,6 +37,7 @@
    Author="Liberzon, A.  and Subramanian, A.  and Pinchback, R.  and Thorvaldsd\'{o}ttir, H.  and Tamayo, P.  and Mesirov, J. P. ",
    Title="{{M}olecular signatures database ({M}{S}ig{D}{B}) 3.0}",
    Journal="Bioinformatics",
+   doi="10.1093/bioinformatics/btr260",
    Year="2011",
    Volume="27",
    Number="12",
@@ -47,6 +49,7 @@
    Author="Ashburner, M.  and Ball, C. A.  and Blake, J. A.  and Botstein, D.  and Butler, H.  and Cherry, J. M.  and Davis, A. P.  and Dolinski, K.  and Dwight, S. S.  and Eppig, J. T.  and Harris, M. A.  and Hill, D. P.  and Issel-Tarver, L.  and Kasarskis, A.  and Lewis, S.  and Matese, J. C.  and Richardson, J. E.  and Ringwald, M.  and Rubin, G. M.  and Sherlock, G. ",
    Title="{{G}ene ontology: tool for the unification of biology. {T}he {G}ene {O}ntology {C}onsortium}",
    Journal="Nat Genet",
+   doi="10.1038/75556",
    Year="2000",
    Volume="25",
    Number="1",
@@ -71,6 +74,7 @@
    Title="{{E}piviz {F}ile {S}erver: {Q}uery, transform and interactively explore data from indexed genomic files}",
    Journal="Bioinformatics",
    Year="2020",
+   doi="10.1093/bioinformatics/btaa591",
    Volume="36",
    Number="18",
    Pages="4682--4690",
@@ -81,6 +85,7 @@
    Author="Subramanian, A.  and Tamayo, P.  and Mootha, V. K.  and Mukherjee, S.  and Ebert, B. L.  and Gillette, M. A.  and Paulovich, A.  and Pomeroy, S. L.  and Golub, T. R.  and Lander, E. S.  and Mesirov, J. P. ",
    Title="{{G}ene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles}",
    Journal="Proc Natl Acad Sci U S A",
+   doi="10.1073/pnas.0506580102",
    Year="2005",
    Volume="102",
    Number="43",
@@ -92,6 +97,7 @@
    Author="Wu, D.  and Smyth, G. K. ",
    Title="{{C}amera: a competitive gene set test accounting for inter-gene correlation}",
    Journal="Nucleic Acids Res",
+   doi="10.1093/nar/gks461",
    Year="2012",
    Volume="40",
    Number="17",
@@ -103,6 +109,7 @@
    Author="Young, M. D.  and Wakefield, M. J.  and Smyth, G. K.  and Oshlack, A. ",
    Title="{{G}ene ontology analysis for {R}{N}{A}-seq: accounting for selection bias}",
    Journal="Genome Biol",
+   doi="10.1186/gb-2010-11-2-r14",
    Year="2010",
    Volume="11",
    Number="2",
@@ -126,6 +133,7 @@
[…]
