### 2023-06-13 #5271 @DavidNickle
I am interested in this but cannot view it.

David

> On Jun 11, 2023, at 2:59 PM, Mark Jensen ***@***.***> wrote:
>
>
> Hey @DavidNickle <https://github.com/DavidNickle> - could you give me a hand with this work <x-msg://2/#5271> for JOSS <https://joss.theoj.org/>?
>
