### 2023-06-26 #5271 @majensen
@svburke - would you be able to review [this work](#5271) for [JOSS](https://joss.theoj.org)? Please let me know - thanks --@majensen
