### 2023-08-20 #5271 @majensen
@bede would you be willing to review [this work](#5271) for [JOSS](https://joss.theoj.org)? I would be grateful for you expertise. Thanks --@majensen
