### 2023-08-20 #5271 @majensen
@betteridiot - would you be able to review [this work](#5271) for [JOSS](https://joss.theoj.org)? Thanks for your consideration --@majensen
