### 2023-08-21 #5271 @arfon
The top 5 most similar previous JOSS papers to this one include:

[GeneNetwork: framework for web-based genetics](https://github.com/openjournals/joss-reviews/issues/25)
Reviewers: `@krother, @agarie`
Similarity score: 0.8190033008132662

[rTASSEL: An R interface to TASSEL for analyzing genomic diversity](https://github.com/openjournals/joss-reviews/issues/4530)
Reviewers: `@tkchafin, @tomsing1`
Similarity score: 0.816965443009842

[A flexible search system for high-accuracy identification of biological entities and molecules](https://github.com/openjournals/joss-reviews/issues/3756)
Reviewers: `@AlexanderPico, @rabdill, @apcamargo`
Similarity score: 0.8117074419436549

[Talisman: a JavaScript archive of fuzzy matching, information retrieval and record linkage building blocks](https://github.com/openjournals/joss-reviews/issues/2405)
Reviewers: `@Fil, @atanikan`
Similarity score: 0.8015637063955903

[geneXplainR: An R interface for the geneXplain platform](https://github.com/openjournals/joss-reviews/issues/412)
Reviewers: `@selbouhaddani`
Similarity score: 0.8012875041284266
