### 2023-08-21 #5271 @betteridiot
If @bede is happy to be the second reviewer, do you still need me?

From: Bede Constantinides
Sent: Monday, August 21, 2023 5:54 AM
To: openjournals/joss-reviews
Cc: Marcus D Sherman; Mention
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: gesel: a JavaScriptpackage for client-side gene set enrichment (Issue #5271)

I'm happy to review this (if the new editors are happy?)
