### 2023-08-22 #5777 @editorialbot
[REVIEW]: gesel: a JavaScript package for client-side gene set enrichment
**Submitting author:** @LTLA (<a href="http://orcid.org/0000-0002-3564-4813">Aaron Lun</a>)
**Repository:** https://github.com/LTLA/gesel.js
**Branch with paper.md** (empty if default branch): submission
**Version:** 0.3.3
