### 2023-08-25 831dadb
func to get embeddings for a species
