### 2023-08-25 622c4d9
Function to get embeddings for each species (#4)
