### 2023-08-25 cb0b0c6
cast tsne coordinates as floats. bump version
