### 2023-08-29 #5777 @LTLA
Thanks to all for getting involved in this. We've added some more functionality to **gesel** in the meantime (including a fun little visualization of all gene sets on a t-SNE), but we'll give everyone a chance to read the current manuscript first.
