### 2023-08-31 a20677d
Renamed fetchEmbeddings function for consistency.
Also fixed up the docs and added some more tests.
