+++ src/index.js
@@ -24,4 +24,4 @@ export { computeEnrichmentCurve } from "./computeEnrichmentCurve.js";
 export { reindexGenesForAllSets } from "./reindexGenesForAllSets.js";
 export { reindexSetsForAllGenes } from "./reindexSetsForAllGenes.js";
 
-export { fetchEmbeddingsForSpecies } from "./fetchEmbeddings.js";
+export { fetchEmbeddings, fetchEmbeddingsForSpecies } from "./fetchEmbeddings.js";
