### 2023-09-02 #5777 @bede
Hi @LTLA, I didn't find any community guidelines. Would you either point me to them or add them? Thank you.
