### 2023-09-04 5c26f4e
Added some contributing guidelines.
