### 2023-09-04 #5777 @LTLA
Hi @bede, done: https://github.com/LTLA/gesel.js/blob/master/CONTRIBUTING.md
