### 2023-09-10 #5777 @majensen
@LTLA - looks like the rendered paper is missing the Subramaniam 2005 ref and the Lun & Manchera 2023 (or is it 2022?). Can you have a look?
