### 2023-09-11 e0cf29c
Updated the kana reference.
