+++ ref.bib
@@ -57,16 +57,17 @@
    Month="May"
 }
 
-@article{lun2022single,
-  author = {Lun, A. and Kancherla, J.},
-  title = {Single-cell data analysis in the browser},
-  elocation-id = {2022.03.02.482701},
-  year = {2022},
-  doi = {10.1101/2022.03.02.482701},
-  publisher = {Cold Spring Harbor Laboratory},
-  URL = {https://www.biorxiv.org/content/early/2022/03/04/2022.03.02.482701},
-  eprint = {https://www.biorxiv.org/content/early/2022/03/04/2022.03.02.482701.full.pdf},
-  journal = {bioRxiv}
+@article{lun2023single,
+   doi = {10.21105/joss.05603},
+   url = {https://doi.org/10.21105/joss.05603},
+   year = {2023},
+   publisher = {The Open Journal},
+   volume = {8},
+   number = {89},
+   pages = {5603},
+   author = {Aaron T. L. Lun and Jayaram Kancherla},
+   title = {Powering single-cell analyses in the browser with WebAssembly},
+   journal = {Journal of Open Source Software} 
 }
 
 @Article{kancherla2020epiviz,
