### 2023-09-11 #5777 @LTLA
@majensen not sure what you mean. The [latest rendering](https://github.com/openjournals/joss-reviews/issues/5777#issuecomment-1703838976) has both references on the last page:

![Screenshot from 2023-09-11 02-42-32](https://github.com/openjournals/joss-reviews/assets/8166669/c69567cc-7c43-4c0a-9a17-aa5b0b5d943d)

I took the liberty of updating Lun and Kancherla reference, given that #5603 has been accepted (yay).
