### 2023-10-02 #5777 @majensen
@arfon Still working on it, but should be complete this week. @LTLA - one request - could you write a few sentences into the paper that describes the companion repo (feedstock) in some more detail? I think it would be helpful to users.
