### 2023-10-12 #5777 @majensen
@LTLA so sorry about my slowness. Maybe you can help speed me up - I've run the code itself (rather than just hitting the site), and it seems to work as advertised. What I want to do is compare its lists with output from e.g. MSigDB. My problem is my lack of experience in the area and the many options for GSEA out there. For example, is the order of genes returned important? (My guess is that has some randomness associated with it, so maybe not.) If some genes are missing in one and present in the other list, is this a problem? Anyway, this is the question I'm trying to formulate. Thoughts?
