### 2023-10-12 #5777 @LTLA
No worries.

> For example, is the order of genes returned important? (My guess is that has some randomness associated with it, so maybe not.)

For the simple hypergeometric test that we use: no, the order doesn't matter. This test only cares about the existence of genes in the set.

There are certainly other tests that care about the ordering of genes, e.g., the original "GSEA" test (as in, the actual test by that name, not the general concept), `limma::geneSetTest`, and so on. However, the relevant order is not that of genes in the pre-defined set, but instead, of the genes that are supplied by the user - typically in order of significance from a differential expression analysis. This allows the test to use more information about the relative importance of genes when considering their enrichment in the set, as opposed to our binary yes/no approach.

Having said that, it is difficult to ask an average user to provide an ordering of genes (at least, if they didn't already do a DE analysis). Hence we use a relatively simple test.

> If some genes are missing in one and present in the other list, is this a problem?

Depends on what lists you're talking about.

I'll consider the most obvious use case where one of the lists is the user-supplied list of genes and the other is a specific gene set. If a gene is missing from one or the other, that's fine and to be expected. The hypergeometric p-value quantifies the significance of the overlap between lists, so if there are lot of genes present in one and not the other, we can expect a low overlap and a large p-value. This will cause the gene set to be lowly ranked in the table of search results.

The less obvious interpretation of your question relates to the "universe" of all possible genes. The previous paragraph assumed that, if we saw any genes in the user's list that were missing from our gene sets, those genes were still valid identifiers that existed somewhere in our annotation for the species of interest. However, if the user's list contains a gene that we don't know about (e.g., transgenes, novel genes), we just ignore that gene as there's not much we can do.

An additional wart to consider is that **gesel**'s universe may not be the same as the user's universe, e.g., if the user used a different reference genome. This can interfere with the calculation of the **gesel** p-values, though they should still be acceptable for ranking gene sets in the search table. (GSEA-related p-values should be taken with a grain of salt anyway, given that they make an often-unreasonable assumption of independence between genes under the null.)
