### 2023-10-23 772ce42
Added some comments about the t-SNE embedding.
