+++ ref.bib
@@ -141,3 +141,11 @@
    Month="Jul"
 }
 
+@article{van2008visualizing,
+  title={Visualizing data using t-SNE.},
+  author={Van der Maaten, Laurens and Hinton, Geoffrey},
+  journal = {Journal of Machine Learning Research},
+  volume={9},
+  number={11},
+  year={2008}
+}
