+++ ref.bib
@@ -142,10 +142,12 @@
 }
 
 @article{van2008visualizing,
-  title={Visualizing data using t-SNE.},
+  title={Visualizing data using t-SNE},
   author={Van der Maaten, Laurens and Hinton, Geoffrey},
   journal = {Journal of Machine Learning Research},
+  year={2008},
   volume={9},
-  number={11},
-  year={2008}
+  number={86},
+  pages={2579--2605},
+  url={http://jmlr.org/papers/v9/vandermaaten08a.html}
 }
