### 2023-12-13 #5777 @majensen
@openjournals/dev : is the below in your domain?

> @Kevin-Mattheus-Moerman Don't know whether this is the most appropriate place to mention this, but I notice that my name doesn't show up properly on the JOSS website (note the lack of capitalization on the L):
>
>
>
> ![Screenshot 2023-12-12 at 2 53 04 PM](https://github.com/openjournals/joss-reviews/assets/8166669/6cfd9e61-18ad-4663-92f6-ea5b7643d608)
>
>
>
> Clicking on the link on my name takes me [here](https://joss.theoj.org/papers/by/Aaron%20T.%20l.%20Lun):
>
>
>
> ![Screenshot 2023-12-12 at 2 54 08 PM](https://github.com/openjournals/joss-reviews/assets/8166669/a774b8ee-e5ed-4007-bd2f-c2dfd625856d)
>
>
>
> This seems to have some implications for indexers, e.g., Google scholar reports:
>
>
>
> ![Screenshot 2023-12-12 at 2 55 32 PM](https://github.com/openjournals/joss-reviews/assets/8166669/5c81dcb1-b683-42b4-9160-456361588427)
>
>
>
> I'm guessing it's my two initials that's causing the system to throw a fit. The same problem with the links is observed for my two other papers on JOSS:
>
>
>
> - https://joss.theoj.org/papers/10.21105/joss.04742
>
> - https://joss.theoj.org/papers/10.21105/joss.05603
>
>
