### 2023-12-13 #5777 @arfon
@LTLA – for your name to appear properly, I think we'll need you to follow this guide to update your `paper.md` to be more explicit about the structure of your name: https://joss.readthedocs.io/en/latest/submitting.html#article-metadata

The error clicking the link on the JOSS site is an unrelated bug.
