### 2023-12-13 #5777 @LTLA
Thanks @arfon, just updated `paper.md` with my full name ("Aaron Tin Long Lun"). Hopefully the heuristics will be smart enough to do the initialization of the multiple middle names properly this time; otherwise, suggestions are welcome.
